release functional v1.0 of eclesia

+++ src/main/java/imdl/eclesia/service/InstrumentoService.java
@@ -17,17 +17,17 @@ public class InstrumentoService {
         this.instrumentoRepository = instrumentoRepository;
     }
 
-    public List<Instrumento> findAll(){
+    public List<Instrumento> findAll() {
         return instrumentoRepository.findAll().stream().map(InstrumentoMapper::entityToDomain).toList();
     }
 
-    public Instrumento findById(Long numero){
+    public Instrumento findById(Long numero) {
         return InstrumentoMapper.entityToDomain(instrumentoRepository.findById(numero)
                 .orElseThrow(() -> new EntityNotFoundException("Instrumento não encontrado.")));
     }
 
-    public Instrumento createInstrumento(InstrumentoInput input){
-        if(instrumentoRepository.existsByNome(input.getNome().toUpperCase()))
+    public Instrumento createInstrumento(InstrumentoInput input) {
+        if (instrumentoRepository.existsByNome(input.getNome().toUpperCase()))
             throw new RogueException("Já existe um instrumento com este nome.");
         Instrumento instrumento = new Instrumento();
         instrumento.setNome(input.getNome().toUpperCase());
@@ -35,9 +35,11 @@ public class InstrumentoService {
         return InstrumentoMapper.entityToDomain(instrumentoRepository.save(InstrumentoMapper.domainToEntity(instrumento)));
     }
 
-    public void deleteInstrumento(Long id){
-        if(!instrumentoRepository.existsById(id))
+    public void deleteInstrumento(Long id) {
+        if (!instrumentoRepository.existsById(id))
             throw new RogueException("Não existe um instrumento com este ID.");
+        if (instrumentoRepository.isInstrumentInUse(instrumentoRepository.findById(id).get()))
+            throw new RogueException("Não é possível excluir este instrumento, pois ele está associado a um levita.");
         instrumentoRepository.deleteById(id);
     }
 }
