+++ src/main/java/imdl/eclesia/service/LevitaService.java
@@ -28,73 +28,79 @@ public class LevitaService {
         this.escalaService = escalaService;
     }
 
-    public List<Levita> findAll(LevitaFilter filter){
+    public List<Levita> findAll(LevitaFilter filter) {
         return levitaRepository.findAll(filter.getNome(), filter.getInstrumento()).stream()
                 .map(LevitaMapper::entityToDomain).sorted(Comparator.comparing(Levita::getNome)).toList();
     }
-    public List<LevitaResumed> findAllResumed(){
+
+    public List<LevitaResumed> findAllResumed() {
         return levitaRepository.findAllResumed().stream().map(LevitaMapper::entityToDomainResumed)
                 .sorted(Comparator.comparing(LevitaResumed::getNome)).toList();
     }
 
-    public List<Levita> findAllById(List<UUID> ids){
+    public List<Levita> findAllById(List<UUID> ids) {
         return levitaRepository.findAllById(ids).stream().map(LevitaMapper::entityToDomain)
                 .sorted(Comparator.comparing(Levita::getNome)).toList();
     }
 
-    public List<LevitaResumed> findAllDisponivel(LocalDate date){
+    public List<LevitaResumed> findAllDisponivel(LocalDate date) {
         return levitaRepository.findAllResumed().stream().map(LevitaMapper::entityToDomainResumed)
                 .filter(levita -> {
-                    if(levita.getAgenda() == null)
+                    if (levita.getAgenda() == null)
                         return true;
                     return !levita.getAgenda().contains(date);
                 }).sorted(Comparator.comparing(LevitaResumed::getNome)).toList();
     }
-    public List<Levita> findAllByInstrument(Long instrumento){
+
+    public List<Levita> findAllByInstrument(Long instrumento) {
         return levitaRepository.findAllByInstrumento(instrumento).stream()
                 .map(LevitaMapper::entityToDomain).sorted(Comparator.comparing(Levita::getNome)).toList();
     }
 
-    public Levita findById(UUID id){
+    public Levita findById(UUID id) {
         return LevitaMapper.entityToDomain(levitaRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Levita não encontrada.")));
     }
-    public Levita create(LevitaInput input){
+
+    public Levita create(LevitaInput input) {
         validateInput(input);
         Levita levita = inputToDomain(input);
         levita.setAgenda(new ArrayList<>());
         return LevitaMapper.entityToDomain(levitaRepository.save(LevitaMapper.domainToEntity(levita)));
     }
-    public Levita update(LevitaInput input){
+
+    public Levita update(LevitaInput input) {
         Levita levita = LevitaMapper.entityToDomain(levitaRepository.findById(input.getId())
                 .orElseThrow(() -> new EntityNotFoundException("Levita não encontrada.")));
-        if(input.getNome() != null)
+        if (input.getNome() != null)
             if (input.getNome().isBlank())
                 throw new RogueException("O nome está vazio.");
             else levita.setNome(input.getNome());
-        if(input.getInstrumentos() != null)
+        if (input.getInstrumentos() != null)
             levita.setInstrumentos(input.getInstrumentos().stream().map(instrumentoService::findById).toList());
-        if(input.getContato() != null)
+        if (input.getContato() != null)
             levita.setContato(input.getContato());
-        if(input.getEmail() != null)
+        if (input.getEmail() != null)
             levita.setEmail(input.getEmail());
-        if(input.getDescricao() != null)
+        if (input.getDescricao() != null)
             levita.setDescricao(input.getDescricao());
         return LevitaMapper.entityToDomain(levitaRepository.save(LevitaMapper.domainToEntity(levita)));
     }
 
-    public void deleteLevita(UUID id){
-        if(escalaService.existsByLevita(id))
-            throw new RogueException("Levita está em uma escala.");
+    public void deleteLevita(UUID id) {
+        if (escalaService.existsByLevita(id))
+            throw new RogueException("Levita não pode ser removido pois já se encontra em uma escala.");
+        if (levitaRepository.existsUserByLevitaId(id))
+            throw new RogueException("Levita não pode ser removido pois já possui um usuário associado.");
         levitaRepository.delete(levitaRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Levita não encontrada.")));
     }
 
-    public Levita addInstrumento(UUID id, Long codInstrumento){ // TODO FIX
+    public Levita addInstrumento(UUID id, Long codInstrumento) { // TODO FIX
         Levita levita = LevitaMapper.entityToDomain(levitaRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Levita não encontrada.")));
         List<Instrumento> instrumentos = levita.getInstrumentos();
-        if(instrumentos.stream().map(Instrumento::getId).toList().contains(codInstrumento))
+        if (instrumentos.stream().map(Instrumento::getId).toList().contains(codInstrumento))
             throw new RogueException("Levita já possui este instrumento.");
         Instrumento instrumento = instrumentoService.findById(codInstrumento);
         List<Instrumento> newInstrumentos = new ArrayList<>();
@@ -108,16 +114,16 @@ public class LevitaService {
         return LevitaMapper.entityToDomain(levitaRepository.save(LevitaMapper.domainToEntity(levita)));
     }
 
-    public Levita removeInstrumento(UUID id, Long instrumento){
+    public Levita removeInstrumento(UUID id, Long instrumento) {
         Levita levita = LevitaMapper.entityToDomain(levitaRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Levita não encontrada.")));
         List<Instrumento> instrumentos = levita.getInstrumentos();
-        if(!instrumentos.stream().map(Instrumento::getId).toList().contains(instrumento))
+        if (!instrumentos.stream().map(Instrumento::getId).toList().contains(instrumento))
             throw new EntityNotFoundException("Levita já não possui este instrumento.");
 
         List<Instrumento> newInstrumentos = new ArrayList<>();
         instrumentos.forEach(i -> {
-            if(!Objects.equals(i.getId(), instrumento))
+            if (!Objects.equals(i.getId(), instrumento))
                 newInstrumentos.add(i);
         });
 
@@ -125,11 +131,11 @@ public class LevitaService {
         return LevitaMapper.entityToDomain(levitaRepository.save(LevitaMapper.domainToEntity(levita)));
     }
 
-    public List<LocalDate> getLevitaAgenda(UUID id){
+    public List<LocalDate> getLevitaAgenda(UUID id) {
         return levitaRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Levita não encontrado.")).getAgenda();
     }
 
-    public Levita updateAgentaFromALevita(UUID id){
+    public Levita updateAgentaFromALevita(UUID id) {
         Levita levita = LevitaMapper.entityToDomain(levitaRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Levita não encontrada.")));
 
@@ -141,26 +147,28 @@ public class LevitaService {
         return levita;
     }
 
-    public Levita setLevitaAgenda(UUID id, List<LocalDate> dates){
+    public Levita setLevitaAgenda(UUID id, List<LocalDate> dates) {
         Levita levita = LevitaMapper.entityToDomain(levitaRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Levita não encontrado.")));
-        if(dates.stream().anyMatch(date -> date.isBefore(LocalDate.now())))
+        if (dates.stream().anyMatch(date -> date.isBefore(LocalDate.now())))
             throw new RogueException("Data já passou.");
+        if (escalaService.existsByLevitaInDates(id, dates))
+            throw new RogueException("Levita está em uma escala em uma das datas selecionadas.");
         levita.setAgenda(dates);
         return LevitaMapper.entityToDomain(levitaRepository.save(LevitaMapper.domainToEntity(levita)));
     }
 
-    private void validateInput(LevitaInput input){
-        if(input.getNome() == null || input.getNome().isBlank())
+    private void validateInput(LevitaInput input) {
+        if (input.getNome() == null || input.getNome().isBlank())
             throw new RogueException("O nome está vazio.");
-        if(input.getInstrumentos() == null || input.getInstrumentos().isEmpty())
+        if (input.getInstrumentos() == null || input.getInstrumentos().isEmpty())
             throw new RogueException("O instrumento está vazio.");
-        if((input.getContato() == null||input.getContato().isBlank())
-                && (input.getEmail() == null ||input.getEmail().isBlank()))
+        if ((input.getContato() == null || input.getContato().isBlank())
+                && (input.getEmail() == null || input.getEmail().isBlank()))
             throw new RogueException("Não há nenhum contato cadastrado.");
     }
 
-    private Levita inputToDomain(LevitaInput input){
+    private Levita inputToDomain(LevitaInput input) {
         Levita levita = new Levita();
         if (input.getId() != null)
             levita.setId(input.getId());
